Добавляет решение задания 'Время разбивать камни'
